I've made all the requested changes except the acceptance test. I'm still working on that.
I have not been able to check the changes with testing yet, because CodeSign wants my Apple Developer account and I don't actually have one. In the past, I've simply logged in with my Apple ID, but that doesn't seem to be working here. I'm not sure why; I was able to run the tests successfully before.