Added ~25 Frames (#25)
* Rebuilding. Tests passed until I tried to change the simulator device, then it wanted a signing code.

* Update README.md

* trying again

* I've made all the requested changes except the acceptance test. I'm still working on that.

I have not been able to check the changes with testing yet, because CodeSign wants my Apple Developer account and I don't actually have one. In the past, I've simply logged in with my Apple ID, but that doesn't seem to be working here. I'm not sure why; I was able to run the tests successfully before.

* Added acceptance tests. All Passing.

* All changes made and tests passing except for the changes to SeriesIndex

* Requested changes implemented. Tests passing.

* oops. Forgot the contributor addition.

* size test lines removed.

* movement index and movement count added as integer frames, tests passing.

* Consolidated ID3FrameTrackPosition and ID3FrameDiscPosition into a single Frame type, ID3FramePartOfTotal. All tests passing.

* renamed file

* updated readme